Add files via upload
In this program, we consider 2 input values, one is a string and the other is an integer 'k'. Here, we will be printing the sub-strings of the string. 'k' is the factor of the length of the string. Using that we will be printing the sub-strings in equal half of the length of the string and the characters in it are not repeated.
Example:
String= 'AABCAAADA'
k = 3
So here, length of String is 9 and k=3 is a factor of 9.
And output will be
AB
CA
AD